Rename "branches" -> "forks" for user repositories.
Since "branch" has a separate meaning from the concept of multiple repositories
per project, it is not appropriate to re-use that term for a clone of the
repository. A "fork" in this case is a copy of one or more branches of a
repository which have a different set of permissions from the original. The idea
is to allow users other than the original creators of the project to commit to a
repository and have it show up on the Drupal site.

Signed-off-by: Dan Hackney <chrono325@gmail.com>